remove tls listener file
